+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Leaflet Changelog
 ## 0.2 (master)
 
  * Added **WMS support** (`TileLayer.WMS`), currently EPSG:3857 only.
+ * Added `TileLayer.Canvas` for easy creation of canvas-based tile layers.
  * `Circle` is now zoom-dependent (with radius in meters); circle of a permanent size is now called `L.CircleMarker`.
  * Added `mouseover` and `mouseout` events to map, markers and paths; added map `mousemove` event.
  * Added `setLatLngs`, `spliceLatLngs`, `addLatLng`, `getLatLngs` methods to polylines and polygons.